feat: move dbState and instance for clustervariable package
