+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/clustervariable/DbClusterVariableState.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.state.variable;
+package io.camunda.zeebe.engine.state.clustervariable;
 
 import io.camunda.zeebe.db.ColumnFamily;
 import io.camunda.zeebe.db.TransactionContext;
